+++ src/main/java/com/hask/hasktask/repository/RefreshTokenRepository.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
     Optional<RefreshToken> findByRefreshToken(String refreshToken);
 
-    Optional<RefreshToken> findByUserId(Integer userId);
-    void deleteByUserId(Integer userId);
+    Optional<RefreshToken> findByUserId(Long userId);
+    void deleteByUserId(Long userId);
 
     /// @Modifying
     // int deleteByUser(User user);
